split login app into login and dashboard
